+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         tools:targetApi="31">
         <activity
             android:name=".ui.serversetting.ServerSettingActivity"
-            android:exported="true" >
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -32,7 +32,7 @@
             android:exported="true" />
         <activity
             android:name=".ui.shopping.ShoppingActivity"
-            android:exported="true"/>
+            android:exported="true" />
     </application>
 
 </manifest>
