--- app/src/main/java/woowacourse/shopping/data/datasource/product/ProductDataSource.kt
@@ -1,13 +1,0 @@
-package woowacourse.shopping.data.datasource.product
-
-import woowacourse.shopping.data.model.Page
-import woowacourse.shopping.data.model.Product
-
-interface ProductDataSource {
-    interface Local
-
-    interface Remote {
-        fun getProductByPage(page: Page): List<Product>
-        fun findProductById(id: Int): Product?
-    }
-}
