--- app/src/main/java/woowacourse/shopping/data/entity/CartEntity.kt
@@ -1,10 +1,0 @@
-package woowacourse.shopping.data.entity
-
-typealias DataCartEntity = CartEntity
-
-class CartEntity(
-    val id: Int,
-    val productId: Int,
-    val count: Int,
-    val checked: Int,
-)
