+++ app/src/main/java/woowacourse/shopping/data/repository/ProductRepositoryImpl.kt
@@ -2,9 +2,8 @@ package woowacourse.shopping.data.repository
 
 import woowacourse.shopping.data.mapper.toData
 import woowacourse.shopping.data.mapper.toDomain
-import woowacourse.shopping.data.service.ProductService
+import woowacourse.shopping.data.service.product.ProductService
 import woowacourse.shopping.domain.model.Product
-import woowacourse.shopping.domain.model.page.Page
 import woowacourse.shopping.domain.repository.ProductRepository
 
 class ProductRepositoryImpl(
@@ -19,11 +18,11 @@ class ProductRepositoryImpl(
         remoteProductDataSource.findProductById(id)?.toDomain()
 
     override fun insertProduct(product: Product) {
-        remoteProductDataSource.addProduct(product.toData())
+        remoteProductDataSource.insertProduct(product.toData())
     }
 
     override fun updateProduct(product: Product) {
-        remoteProductDataSource.adjustProduct(product.toData())
+        remoteProductDataSource.updateProduct(product.toData())
     }
 
     override fun deleteProduct(product: Product) {
