--- app/src/main/java/woowacourse/shopping/util/inject/DaoInject.kt
@@ -1,8 +1,0 @@
-package woowacourse.shopping.util.inject
-
-import woowacourse.shopping.data.database.ShoppingDatabase
-import woowacourse.shopping.data.database.dao.recentproduct.RecentProductDao
-import woowacourse.shopping.data.database.dao.recentproduct.RecentProductDaoImpl
-
-fun injectRecentProductDao(database: ShoppingDatabase): RecentProductDao =
-    RecentProductDaoImpl(database)
