--- domain/src/main/java/woowacourse/shopping/domain/model/CartEntity.kt
@@ -1,10 +1,0 @@
-package woowacourse.shopping.domain.model
-
-typealias DomainCartEntity = CartEntity
-
-class CartEntity(
-    val id: Int,
-    val productId: Int,
-    val count: Int,
-    val checked: Boolean,
-)
